[move] restructure to mvc model
MVC 모델에 따라서 패키지 재배치
