+++ .gitignore
@@ -118,7 +118,7 @@ Icon
 .VolumeIcon.icns
 .com.apple.timemachine.donotpresent
 
-# Directories potentially created on remote AFP share
+# Directories potentially created on data AFP share
 .AppleDB
 .AppleDesktop
 Network Trash Folder
