+++ app/src/main/java/org/android/go/sopt/data/local/mockPlayListTitle.kt
@@ -1,5 +1,7 @@
 package org.android.go.sopt.presentation.playlist
 
+import org.android.go.sopt.data.local.PlaylistTitle
+
 val mockPlayListTitle = listOf<PlaylistTitle>(
     PlaylistTitle(0, "상호의 플레이리스트")
 )
