+++ app/src/main/java/org/android/go/sopt/data/remote/SignUpRequestDTO.kt
@@ -1,12 +1,16 @@
-package org.android.go.sopt.remote.auth
+package org.android.go.sopt.data.remote
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class LoginRequestDTO(
+data class SignUpRequestDTO(
     @SerialName("id")
     val id: String,
     @SerialName("password")
     val password: String,
+    @SerialName("name")
+    val name: String,
+    @SerialName("skill")
+    val skill: String,
 )
