+++ app/src/main/java/org/android/go/sopt/presentation/auth/SignUpActivity.kt
@@ -8,10 +8,10 @@ import android.os.Bundle
 import android.view.inputmethod.InputMethodManager
 import androidx.core.widget.doAfterTextChanged
 import org.android.go.sopt.R
-import org.android.go.sopt.remote.auth.SignUpRequestDTO
-import org.android.go.sopt.remote.auth.SignUpResponseDTO
+import org.android.go.sopt.data.remote.SignUpRequestDTO
+import org.android.go.sopt.data.remote.SignUpResponseDTO
 import org.android.go.sopt.databinding.ActivitySignUpBinding
-import org.android.go.sopt.remote.auth.AuthServicePool.authService
+import org.android.go.sopt.module.AuthServicePool.authService
 import org.android.go.sopt.util.KeyboardVisibilityUtils
 import org.android.go.sopt.util.makeSnackBar
 import retrofit2.Call
