+++ app/src/main/java/org/android/go/sopt/presentation/playlist/PlaylistTitleAdapter.kt
@@ -3,6 +3,7 @@ package org.android.go.sopt.presentation.playlist
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
+import org.android.go.sopt.data.local.PlaylistTitle
 import org.android.go.sopt.databinding.ItemPlaylistTitleBinding
 import org.android.go.sopt.util.ItemDiffCallback
